requirements.txt
deep_translator
asyncio
fastapi
websocket
groq
supabase
python-dotenv
pip install python-multipart
redis==6.4.0
pydub
uvicorn==0.37.0
pydantic[email]
